+++ BE/src/main/java/com/squadmap/common/auth/application/dto/LoginInfo.java
@@ -1,0 +1,15 @@
+package com.squadmap.common.auth.application.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class LoginInfo {
+
+    private final Long memberId;
+    private final String nickname;
+    private final String profileImage;
+    private final Tokens tokens;
+
+}
