+++ BE/src/main/java/com/squadmap/common/auth/application/dto/github/GithubUserInfo.java
@@ -1,0 +1,11 @@
+package com.squadmap.common.auth.application.dto.github;
+
+import lombok.Getter;
+
+@Getter
+public class GithubUserInfo {
+
+    private String login;
+    private String avatarUrl;
+
+}
