+++ BE/src/main/java/com/squadmap/common/auth/application/dto/naver/NaverUserInfo.java
@@ -1,0 +1,11 @@
+package com.squadmap.common.auth.application.dto.naver;
+
+import lombok.Getter;
+
+@Getter
+public class NaverUserInfo {
+
+    private String nickname;
+    private String email;
+    private String profileImage;
+}
