+++ BE/src/main/java/com/squadmap/common/auth/application/dto/naver/NaverUserProfileResponse.java
@@ -1,0 +1,9 @@
+package com.squadmap.common.auth.application.dto.naver;
+
+import lombok.Getter;
+
+@Getter
+public class NaverUserProfileResponse {
+
+    private NaverUserInfo response;
+}
