+++ BE/src/main/java/com/squadmap/common/auth/ui/dto/LoginResponse.java
@@ -1,0 +1,17 @@
+package com.squadmap.common.auth.ui.dto;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+@EqualsAndHashCode
+public class LoginResponse {
+
+    private final String accessToken;
+    private final String refreshToken;
+    private final String userNickname;
+    private final String userProfileUrl;
+
+}
