+++ BE/src/main/java/com/squadmap/common/config/JwtConfig.java
@@ -1,0 +1,16 @@
+package com.squadmap.common.config;
+
+import com.squadmap.common.properties.JwtProperties;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@EnableConfigurationProperties(JwtProperties.class)
+@PropertySource("classpath:properties/auth.yml")
+@RequiredArgsConstructor
+public class JwtConfig {
+
+    private final JwtProperties jwtProperties;
+}
