+++ BE/src/main/java/com/squadmap/common/config/OauthConfig.java
@@ -1,0 +1,17 @@
+package com.squadmap.common.config;
+
+import com.squadmap.common.properties.OauthProperties;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@EnableConfigurationProperties(OauthProperties.class)
+@PropertySource("classpath:properties/auth.yml")
+@RequiredArgsConstructor
+public class OauthConfig {
+
+    private final OauthProperties oauthProperties;
+
+}
