--- BE/src/main/java/com/squadmap/member/application/OauthProvider.java
@@ -1,10 +1,0 @@
-package com.squadmap.member.application;
-
-import com.squadmap.member.application.dto.MemberInfo;
-import com.squadmap.member.config.properties.OauthProperties;
-
-public interface OauthProvider {
-
-    MemberInfo oauth(String code, String state, OauthProperties.OauthProperty oauthProperty);
-
-}
