--- BE/src/main/java/com/squadmap/member/application/dto/MemberInfo.java
@@ -1,14 +1,0 @@
-package com.squadmap.member.application.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class MemberInfo {
-
-    private final String nickname;
-    private final String avatarUrl;
-    private final String email;
-
-}
