--- BE/src/main/java/com/squadmap/member/application/dto/Tokens.java
@@ -1,12 +1,0 @@
-package com.squadmap.member.application.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class Tokens {
-
-    private final String accessToken;
-    private final String refreshToken;
-}
