--- BE/src/main/java/com/squadmap/member/application/dto/github/GithubUserInfo.java
@@ -1,11 +1,0 @@
-package com.squadmap.member.application.dto.github;
-
-import lombok.Getter;
-
-@Getter
-public class GithubUserInfo {
-
-    private String login;
-    private String avatarUrl;
-
-}
