--- BE/src/main/java/com/squadmap/member/application/dto/naver/NaverUserInfo.java
@@ -1,11 +1,0 @@
-package com.squadmap.member.application.dto.naver;
-
-import lombok.Getter;
-
-@Getter
-public class NaverUserInfo {
-
-    private String nickname;
-    private String email;
-    private String profileImage;
-}
