--- BE/src/main/java/com/squadmap/member/application/dto/naver/NaverUserProfileResponse.java
@@ -1,9 +1,0 @@
-package com.squadmap.member.application.dto.naver;
-
-import lombok.Getter;
-
-@Getter
-public class NaverUserProfileResponse {
-
-    private NaverUserInfo response;
-}
