--- BE/src/main/java/com/squadmap/member/config/properties/JwtProperties.java
@@ -1,22 +1,0 @@
-package com.squadmap.member.config.properties;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
-
-
-@ConfigurationProperties(prefix = "jwt")
-@ConstructorBinding
-@RequiredArgsConstructor
-@Getter
-public class JwtProperties {
-
-    private final String secretKey;
-    private final Long accessExpireTime;
-    private final Long refreshExpireTime;
-    private final String accessTokenSubject;
-    private final String refreshTokenSubject;
-    private final String issuer;
-    private final String tokenType;
-}
