--- BE/src/main/java/com/squadmap/member/ui/login/dto/LoginRequest.java
@@ -1,16 +1,0 @@
-package com.squadmap.member.ui.login.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class LoginRequest {
-
-    private String code;
-    private String state;
-
-}
