--- BE/src/main/java/com/squadmap/member/ui/login/dto/LoginResponse.java
@@ -1,17 +1,0 @@
-package com.squadmap.member.ui.login.dto;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-@EqualsAndHashCode
-public class LoginResponse {
-
-    private final String accessToken;
-    private final String refreshToken;
-    private final String userNickname;
-    private final String userProfileUrl;
-
-}
